- renamed FormCrawler to DefaultUploader - moved inner classes to util package - extracted static methods to final class SyncHelper
